+++ mcp_tools/gateway_server/tools/index.ts
@@ -2,3 +2,4 @@ export { researchTools } from "./researchTools.js";
 export { narrativeTools } from "./narrativeTools.js";
 export { verificationTools } from "./verificationTools.js";
 export { knowledgeTools } from "./knowledgeTools.js";
+export { documentTools } from "./documentTools.js";